requirements.txt
albumentations==1.3.0
imageio==2.28.1
numpy==1.24.3
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
Pillow==9.5.0
Pillow==9.5.0
scipy==1.10.1
streamlit==1.22.0
torch==2.0.1
torchvision==0.15.2
